refactor: 매칭 요청 취소 시 유효성 검증 추가
- TUTOR_WAITING & TUTEE_WAITING일 경우에만 완전 삭제 가능하도록 유효성 검증 추가

+++ server/src/main/java/com/mainproject/server/tutoring/service/TutoringService.java
@@ -119,8 +119,15 @@ public class TutoringService {
     }
 
     public void deleteTutoring(Long tutoringId) {
-        messageService.deleteMessageRoomTutoringId(tutoringId);
-        tutoringRepository.delete(verifiedTutoring(tutoringId));
+        Tutoring tutoring = verifiedTutoring(tutoringId);
+        if (tutoring.getTutoringStatus().equals(TutoringStatus.TUTEE_WAITING) ||
+                tutoring.getTutoringStatus().equals(TutoringStatus.TUTOR_WAITING)
+        ) {
+            messageService.deleteMessageRoomTutoringId(tutoringId);
+            tutoringRepository.delete(tutoring);
+        } else {
+            throw new ServiceLogicException(ErrorCode.TUTORING_STATUS_BAD_REQUEST);
+        }
     }
 
     /* 검증 및 유틸 로직 */
